+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Increased size of Detector textbox (way2muchnoise)
 - Autocrafting bugfixes (way2muchnoise)
 - Fixed stack upgrades not working in exporter when stack size is 16 (way2muchnoise)
+- Fixed crash when rotating External Storage (raoulvdberge)
 
 ### 1.2.18
 - Performance improvements with oredict autocrafting (way2muchnoise)